:art: refactor: Moved to appropriate namespaces
