+++ OutOfTimePrototype/Services/General/Interfaces/ICampusBuildingService.cs
@@ -2,7 +2,7 @@ using OutOfTimePrototype.DAL.Models;
 using OutOfTimePrototype.DTO;
 using OutOfTimePrototype.Utilities;
 
-namespace OutOfTimePrototype.Services;
+namespace OutOfTimePrototype.Services.General.Interfaces;
 
 public interface ICampusBuildingService
 {
